+++ extensions/prompt-stash/CHANGELOG.md
@@ -1,3 +1,7 @@
 # PromptStash Changelog
 
+## [Feature] - {PR_MERGE_DATE}
+
+- Added an edit prompt action to the prompt list.
+
 ## [Initial Version] - 2024-12-17